+++ tests/test_utils.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+from django.core.urlresolvers import reverse
 from django.test.testcases import TestCase
 
 from cruds import utils
@@ -28,4 +29,17 @@ class TestUtils(TestCase):
         self.assertEqual(
             utils.crud_permission_name(Author, utils.ACTION_LIST),
             'testapp.list_author'
+        )
+
+    def test_crud_url_for_instance(self):
+        instance = Author.objects.create(name='foo')
+        self.assertEqual(
+            utils.crud_url(instance, utils.ACTION_DETAIL),
+            '/testapp/author/1/'
+        )
+
+    def test_crud_url_for_model(self):
+        self.assertEqual(
+            utils.crud_url(Author, utils.ACTION_LIST),
+            '/testapp/author/'
         )